+++ CHANGES.md
@@ -3,3 +3,7 @@
 ## 0.0.1 2015-03-14 Jason Antman <jason@jasonantman.com>
 
 * Initial release with bitlbee, http_get and virtualenv types
+
+## 0.0.2 2015-03-14 Jason Antman <jason@jasonantman.com>
+
+* Minor bugfix in Serverspec::Types::Virtualenv#virtualenv?